Add State.swift
